refactor: unify JWT settings validation across all hosts
- Move JwtSettings from Backend.Web.Configuration to Contracts.Common.Auth
  (shared by Backend, split-host UI, and single-host UI)
- Add ConfigureBlazorJwtBearerOptions (IConfigureNamedOptions) for split-host
  mirroring backend's ConfigureJwtBearerOptions pattern
- Split-host now uses IOptions<JwtSettings> with ValidateDataAnnotations()
  and ValidateOnStart() — matching backend's validation behavior
- All hosts read JWT config from SecuritySettings:JwtSettings
