+++ src/AditiKraft.Krafter.Backend/Features/Auth/RefreshToken.cs
@@ -2,7 +2,7 @@ using System.IdentityModel.Tokens.Jwt;
 using System.Security.Claims;
 using System.Text;
 using AditiKraft.Krafter.Backend.Web;
-using AditiKraft.Krafter.Backend.Web.Configuration;
+using AditiKraft.Krafter.Contracts.Common.Auth;
 using AditiKraft.Krafter.Backend.Features.Auth.Common;
 using AditiKraft.Krafter.Backend.Features.Users.Common;
 using AditiKraft.Krafter.Backend.Infrastructure.Persistence;
